# backend/services/saved-service/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Inflap Saved Service API
  version: 1.0.0
  description: |
    Owner-scoped HTTP contract for Saved items and private collections.
    The authenticated owner is derived only from the verified session; request
    paths, query parameters, and bodies never accept an owner identifier.
    All personal responses are private and non-cacheable.
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/users/me/saved-items:
    parameters:
      - $ref: '#/components/parameters/ClientPlatformHeader'
      - $ref: '#/components/parameters/AppBuildHeader'
      - $ref: '#/components/parameters/AcceptLanguage'
    get:
      tags: [Saved items]
      summary: List saved items
      description: |
        Returns an owner-scoped cursor page ordered by server state. A
        collection_id filter and uncollected=true are mutually exclusive. A
        foreign or deleted collection has the same neutral not-found response.
        Exact aggregate and per-category counters are intentionally absent.
      operationId: listSavedItems
      parameters:
        - $ref: '#/components/parameters/EntityTypeQuery'
        - $ref: '#/components/parameters/CollectionIdQuery'
        - $ref: '#/components/parameters/UncollectedQuery'
        - $ref: '#/components/parameters/CursorQuery'
        - $ref: '#/components/parameters/PageLimitQuery'
      responses:
        '200':
          $ref: '#/components/responses/SavedItemsPageResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
...
  /v1/users/me/saved-operations/{operationId}:
    parameters:
      - $ref: '#/components/parameters/ClientPlatformHeader'
      - $ref: '#/components/parameters/AppBuildHeader'
      - $ref: '#/components/parameters/OperationIdPath'
      - $ref: '#/components/parameters/AcceptLanguage'
    get:
      tags: [Saved operations]
      summary: Get operation status
      description: |
        Reads an operation only for its originating subject and login-session
        generation. It never starts or replays a mutation. Foreign, expired-out
        of retention, and unknown identifiers return the same neutral NOT_FOUND.
      operationId: getSavedOperation
      responses:
        '200':
          $ref: '#/components/responses/OperationStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
  /v1/users/me/saved-collections:
    parameters:
      - $ref: '#/components/parameters/ClientPlatformHeader'
      - $ref: '#/components/parameters/AppBuildHeader'
      - $ref: '#/components/parameters/AcceptLanguage'
    post:
      tags: [Saved collections]
      summary: Create a private collection
      operationId: createSavedCollection
      parameters:
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - $ref: '#/components/parameters/SavedSourceSurfaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/MutationTerminalResponse'
        '202':
          $ref: '#/components/responses/MutationPendingResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
    get:
      tags: [Saved collections]
      summary: List all active private collections
      description: |
        Returns one bounded owner-scoped list of at most 200 active collections,
        ordered by organized_at descending and then collection ID descending.
        There is no server pagination or collection-title search contract.
      operationId: listSavedCollections
      responses:
        '200':
          $ref: '#/components/responses/SavedCollectionsListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
  /v1/users/me/saved-collections/{collectionId}:
    parameters:
      - $ref: '#/components/parameters/ClientPlatformHeader'
      - $ref: '#/components/parameters/AppBuildHeader'
      - $ref: '#/components/parameters/CollectionIdPath'
      - $ref: '#/components/parameters/AcceptLanguage'
    get:
      tags: [Saved collections]
      summary: Get private collection metadata
      description: |
        Returns metadata only. Collection cards are read through the Saved items
        list with collection_id in its owner-scoped query.
      operationId: getSavedCollection
      responses:
        '200':
          $ref: '#/components/responses/SavedCollectionDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
    patch:
      tags: [Saved collections]
      summary: Rename a private collection
      description: Changes only the title and requires the current metadata version.
      operationId: renameSavedCollection
      parameters:
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - $ref: '#/components/parameters/SavedSourceSurfaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameCollectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/MutationTerminalResponse'
        '202':
          $ref: '#/components/responses/MutationPendingResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/UnprocessableEntityError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
    delete:
      tags: [Saved collections]
      summary: Delete a private collection
      description: |
        Requires lifecycle and metadata versions, but no item-count version.
        Deletion hides the collection and its effective memberships atomically;
        it never unsaves the underlying relationships.
      operationId: deleteSavedCollection
      parameters:
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
        - $ref: '#/components/parameters/SavedSourceSurfaceHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteCollectionRequest'
      responses:
        '200':
          $ref: '#/components/responses/MutationTerminalResponse'
        '202':
          $ref: '#/components/responses/MutationPendingResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
        '503':
          $ref: '#/components/responses/ServiceUnavailableError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified user access token with a current login-session generation.
  parameters:
    ClientPlatformHeader:
      name: X-Client-Platform
      in: header
      required: false
      description: |
        Bounded mobile platform input for owner-sticky product rollout. It is
        not an authentication or authorization claim. Missing or invalid input
        disables new expansion but never blocks canonical reads or reduction.
      schema:
        type: string
        enum: [android, ios]
    AppBuildHeader:
      name: X-App-Build
      in: header
      required: false
      description: |
        Positive decimal store build used only by product rollout minimum-build
        rules. It grants no data access. Missing or invalid input disables new
        expansion but never blocks canonical reads or reduction.
      schema:
        type: integer
        format: uint64
        minimum: 1
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: false
      description: |
        Requested Saved presentation locale. The service normalizes the value to
        one effective EN, RU, or KK locale and returns it in Content-Language.
      schema:
        $ref: '#/components/schemas/AcceptLanguageValue'
...
    OperationIdPath:
      name: operationId
      in: path
      required: true
      description: Client-generated UUIDv4 operation identity.
      schema:
        $ref: '#/components/schemas/OperationId'
    CollectionIdPath:
      name: collectionId
      in: path
      required: true
      description: Opaque owner-scoped collection UUID; possession grants no access.
      schema:
        $ref: '#/components/schemas/CollectionId'
    OperationIdHeader:
      name: Operation-Id
      in: header
      required: true
      description: |
        Cryptographically random client-generated UUIDv4 identifying exactly one
        mutation attempt within the current login-session generation.
      schema:
        $ref: '#/components/schemas/OperationId'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Base64url key generated from at least 16 random bytes. It must be
        cryptographically independent of Operation-Id and must never be reused
        for a different resource or command.
      schema:
        type: string
        minLength: 22
        maxLength: 128
        pattern: '^[A-Za-z0-9_-]{22,128}$'
        x-min-entropy-bits: 128
    SavedSourceSurfaceHeader:
      name: Saved-Source-Surface
      in: header
      required: false
      description: |
        Bounded first-party UI attribution for the accepted attempt. It is not
        part of mutation semantics or idempotency matching. Missing values are
        normalized to UNKNOWN; arbitrary analytics labels are rejected.
      schema:
        type: string
        default: UNKNOWN
        enum: [UNKNOWN, CARD, DETAIL, SAVED_ALL, SAVED_COLLECTION]
    EntityTypeQuery:
      name: type
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/EntityType'
    CollectionIdQuery:
      name: collection_id
      in: query
      required: false
      description: Mutually exclusive with uncollected=true.
      schema:
        $ref: '#/components/schemas/CollectionId'
    UncollectedQuery:
      name: uncollected
      in: query
      required: false
      description: When true, mutually exclusive with collection_id.
      schema:
        type: boolean
        default: false
    CursorQuery:
      name: cursor
      in: query
      required: false
      description: Opaque owner-, scope-, filter-, and locale-bound continuation token.
      schema:
        $ref: '#/components/schemas/OpaqueCursor'
    PageLimitQuery:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 30
...
  responses:
    MutationTerminalResponse:
      description: The accepted mutation has reached an immutable terminal result.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TerminalOperationResult'
...
    SavedItemsPageResponse:
      description: Owner-scoped Saved page without aggregate counters.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SavedItemsPage'
...
    OperationStatusResponse:
      description: Current immutable operation status plus an optional separately fresh snapshot.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationResult'
...
    BadRequestError:
      description: Request syntax, encoding, cursor, filter, or validation failure.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnauthorizedError:
      description: The verified authenticated session is absent or invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        WWW-Authenticate:
          $ref: '#/components/headers/WWWAuthenticate'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ForbiddenError:
      description: Personal-data access is blocked by the shared platform policy.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Neutral not-found response that does not reveal foreign personal resources.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequestsError:
      description: Subject, IP, operation, or owner-quota rate control rejected the request.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServiceUnavailableError:
      description: Saved or a required source dependency is temporarily unavailable.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateCacheControl'
        Pragma:
          $ref: '#/components/headers/PragmaNoCache'
        Content-Language:
          $ref: '#/components/headers/ContentLanguage'
        Vary:
          $ref: '#/components/headers/VaryPersonal'
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    EntityType:
      type: string
      enum: [ATTRACTION, ACTIVITY, USER, POST]
...
    CreateCollectionRequest:
      type: object
      additionalProperties: false
      required: [client_creation_id, title]
      properties:
        client_creation_id:
          $ref: '#/components/schemas/ClientCreationId'
          description: Immutable identity for this logical collection creation.
        title:
          $ref: '#/components/schemas/CollectionTitle'
...